ID-823 create paymentIntentDetails fragment
