+++ Core/GraphQL/GraphQLFiles/sell_createPaymentIntent.graphql
@@ -1,21 +1,5 @@
 mutation createPaymentIntent($input: PaymentIntentInput!) {
   createPaymentIntent(input: $input) {
-    id
-    orderType
-    orderId
-    status
-    paymentSource {
-      ...paymentSourceDetails
-    }
-    amount
-    currency
-    livemode
-    cancellationReason
-    savePaymentSource
-    nextAction {
-      type
-      url
-    }
-    lastPaymentError
+    ...paymentIntentDetails
   }
 }
