+++ Core/GraphQL/GraphQLFiles/sell_fragment_paymentIntentDetails.graphql
@@ -1,0 +1,19 @@
+fragment paymentIntentDetails on PaymentIntent {
+  id
+  orderType
+  orderId
+  status
+  paymentSource {
+    ...paymentSourceDetails
+  }
+  amount
+  currency
+  livemode
+  cancellationReason
+  savePaymentSource
+  nextAction {
+    type
+    url
+  }
+  lastPaymentError
+}
