+++ Core/GraphQL/GraphQLFiles/sell_getMyPaymentIntent.graphql
@@ -1,21 +1,5 @@
 query getMyPaymentIntent($id: ID!) {
   getMyPaymentIntent(id: $id) {
-    id
-    orderType
-    orderId
-    status
-    paymentSource {
-      ...paymentSourceDetails
-    }
-    amount
-    currency
-    livemode
-    cancellationReason
-    savePaymentSource
-    nextAction {
-      type
-      url
-    }
-    lastPaymentError
+    ...paymentIntentDetails
   }
 }
